% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/generateSimpleDataset.R
\name{generateSimpleDataset}
\alias{generateSimpleDataset}
\title{generateSimpleDataset}
\usage{
generateSimpleDataset(n)
}
\arguments{
\item{n}{taille du jeu de données}
}
\description{
Cette fonction génère un jeu de donnée simple
}
\examples{
generateSimpleDataset(100)
}
\keyword{dataset}
\keyword{generation}

