% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.accuracy}
\alias{sc.accuracy}
\title{sc.accuracy}
\usage{
sc.accuracy(M)
}
\arguments{
\item{M}{matrice de confusion}
}
\description{
Cette fonction calcule l'accuracy à partir d'une matrice de confusion
}
\examples{
sc.accuracy(matriceDeConfusion)
}
\keyword{accuracy}

