% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.erreur}
\alias{sc.erreur}
\title{sc.erreur}
\usage{
sc.erreur(M)
}
\arguments{
\item{M}{matrice de confusion}
}
\description{
Cette fonction calcule le taux d'erreur à partir d'une matrice de confusion
}
\examples{
sc.erreur(matriceDeConfusion)
}
\keyword{error}

