% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.evaluation}
\alias{sc.evaluation}
\title{sc.evaluation}
\usage{
sc.evaluation(preds, labels)
}
\arguments{
\item{preds}{prédictions
labels réalité}
}
\description{
Cette fonction renvoie une liste de l'accuracy, la sensibilité, la spécificité, la précision et le taux d'erreur calculés à partir de prédictions
}
\examples{
sc.evaluation(predictions,labels)
}
\keyword{accuracy}
\keyword{confusion}
\keyword{error}
\keyword{matrix}
\keyword{precision}
\keyword{sensitivity}
\keyword{specificity}

