% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.generateDifficultDataset.R
\name{sc.generateDifficultDataset}
\alias{sc.generateDifficultDataset}
\title{sc.generateDifficultDataset}
\usage{
sc.generateDifficultDataset(n)
}
\arguments{
\item{n}{taille du jeu de données}
}
\description{
Cette fonction génère un jeu de donnée 'Difficult'
}
\examples{
sc.generateDifficultDataset(100)
}
\keyword{dataset}
\keyword{dimensions}
\keyword{generation}

