% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.generateDifficultDatasetAlt.R
\name{sc.generateDifficultDatasetAlt}
\alias{sc.generateDifficultDatasetAlt}
\title{sc.generateDifficultDatasetAlt}
\usage{
sc.generateDifficultDatasetAlt(d, n)
}
\arguments{
\item{d}{nombre de dimensions
n taille du jeu de données}
}
\description{
Cette fonction génère un jeu de donnée 'Difficult' en plusieurs dimensions
}
\examples{
sc.generateDifficultDatasetAlt(100,30)
}
\keyword{dataset}
\keyword{dimensions}
\keyword{generation}

