% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.generateMultivariateGaussian.R
\name{sc.generateMultivariateGaussian}
\alias{sc.generateMultivariateGaussian}
\title{sc.generateMultivariateGaussian}
\usage{
sc.generateMultivariateGaussian(n)
}
\arguments{
\item{n}{nombre de vecteurs}
}
\description{
Cette fonction génère des vecteurs dans R² selon une loi normale multivariée
}
\examples{
sc.generateMultivariateGaussian(100)
}
\keyword{dataset}
\keyword{generation}
\keyword{multivariate}

