% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.generateSimpleDataset.R
\name{sc.generateSimpleDataset}
\alias{sc.generateSimpleDataset}
\title{sc.generateSimpleDataset}
\usage{
sc.generateSimpleDataset(n)
}
\arguments{
\item{n}{taille du jeu de données}
}
\description{
Cette fonction génère un jeu de donnée simple
}
\examples{
sc.generateSimpleDataset(100)
}
\keyword{dataset}
\keyword{generation}

