% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.holdOut.R
\name{sc.holdOut}
\alias{sc.holdOut}
\title{sc.holdOut}
\usage{
sc.holdOut(f, d_train, d_test, pred = TRUE)
}
\arguments{
\item{f}{fonction de prediction
d_train jeu d'apprentissage
d_test jeu de test
pred(optionnal, default=TRUE) precise si l'on souhaite des prédictions ou des scores}
}
\description{
Cette fonction opère un hold-out
}
\examples{
sc.holdOut(sc.mediatorHyperplane,d_train,d_test)
}
\keyword{error}

