% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.mediator.demo.R
\name{sc.mediator.demo}
\alias{sc.mediator.demo}
\title{sc.mediator.demo}
\usage{
sc.mediator.demo(d)
}
\arguments{
\item{d}{jeu de données}
}
\description{
Cette fonction génère un graphique à partir d'un jeu de données et représente les centre d'inertie ainsi que l'hyperplan médiateur
}
\examples{
sc.mediator.demo(sc.generateSimpleDataset(100))
}
\keyword{dataset}
\keyword{demo}
\keyword{display}
\keyword{hyperplane}
\keyword{mediator}
\keyword{plot}

