% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.mediatorHyperplane.R
\name{sc.mediatorHyperplane}
\alias{sc.mediatorHyperplane}
\title{sc.mediatorHyperplane}
\usage{
sc.mediatorHyperplane(d)
}
\arguments{
\item{d}{jeu de données}
}
\description{
Cette fonction produit une liste contenant les vecteurs w et b ainsi que deux fonctions de prédiction nommées pred et f prenant en entrée un ensemble de vecteurs.
}
\examples{
sc.mediatorHyperplane(sc.generateDifficultDatasetAlt(100,30))
}
\keyword{b}
\keyword{f}
\keyword{hyperplane}
\keyword{mediator}
\keyword{pred}
\keyword{w}

