% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.plot.dataset.R
\name{sc.plot.dataset}
\alias{sc.plot.dataset}
\title{sc.plot.dataset}
\usage{
sc.plot.dataset(d)
}
\arguments{
\item{d}{jeu de données}
}
\description{
Cette fonction génère un graphique à partir d'un jeu de données
}
\examples{
sc.plot.dataset(sc.generateSimpleDataset(100))
}
\keyword{dataset}
\keyword{display}
\keyword{plot}

