% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.precision}
\alias{sc.precision}
\title{sc.precision}
\usage{
sc.precision(M)
}
\arguments{
\item{M}{matrice de confusion}
}
\description{
Cette fonction calcule la précision à partir d'une matrice de confusion
}
\examples{
sc.precision(matriceDeConfusion)
}
\keyword{precision}

