% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.sensitivity}
\alias{sc.sensitivity}
\title{sc.sensitivity}
\usage{
sc.sensitivity(M)
}
\arguments{
\item{M}{matrice de confusion}
}
\description{
Cette fonction calcule la sensibilité à partir d'une matrice de confusion
}
\examples{
sc.sensitivity(matriceDeConfusion)
}
\keyword{sensitivity}

