% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/sc.calculs.R
\name{sc.specificity}
\alias{sc.specificity}
\title{sc.specificity}
\usage{
sc.specificity(M)
}
\arguments{
\item{M}{matrice de confusion}
}
\description{
Cette fonction calcule la spécificité à partir d'une matrice de confusion
}
\examples{
sc.specificity(matriceDeConfusion)
}
\keyword{specificity}

